requirements.txt
fastapi
uvicorn[standard]
opencv-python
scipy
cvzone
numpy
mediapipe
sqlmodel
sqlalchemy
psycopg2-binary
python-multipart
bcrypt==4.0.0
python-jose[cryptography]==3.3.0
typing-extensions
